requirements.txt
# Core Flask dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Database
pymongo==4.6.0

# Forms and Security (Optional - will work without these)
Flask-WTF==1.2.1
WTForms==3.1.1

# Environment management
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0
waitress==2.1.2

# Development and Testing (Optional)
pytest==7.4.3
pytest-flask==1.3.0
flake8==6.0.0